mysite/requirements.txt
asgiref==3.7.2
autopep8==2.3.1
Django==5.0
djangorestframework==3.15.2
pycodestyle==2.12.1
sqlparse==0.4.4
typing_extensions==4.9.0
